Feature/documentation

+++ core/src/test/scala/co.blocke.scalajack/json/test.misc/ComplexRelationships.scala
@@ -39,7 +39,7 @@ class ComplexRelationships extends FunSpec with GivenWhenThen with BeforeAndAfte
         inst.asInstanceOf[Dog[_]].kind.asInstanceOf[Dog[_]].kind.getClass
       }
       // Loss of fidelity on read: Long -> Byte
-      assertResult(classOf[java.lang.Byte]) {
+      assertResult(classOf[java.lang.Integer]) {
         val found = sj.read[Pet](js)
         found.asInstanceOf[Dog[_]].kind.asInstanceOf[Dog[_]].kind.getClass
       }
