+++ core/src/test/scala/co.blocke.scalajack/json/test.primitives/ValueClassPrim.scala
@@ -160,7 +160,7 @@ class ValueClassPrim() extends FunSpec with Matchers {
         assertResult("""25""") { js }
         assertResult((inst, true)) {
           val r = sj.read[VCNumber](js)
-          (r, r.vc.isInstanceOf[Byte])
+          (r, r.vc.isInstanceOf[Integer])
         }
       }
     }
